requirements.txt
asgiref==3.3.4
Django==3.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
